chore: (#281) Code 파일 분리
