+++ simtong-domain/src/main/kotlin/team/comit/simtong/domain/auth/model/AuthCode.kt
@@ -1,6 +1,6 @@
 package team.comit.simtong.domain.auth.model
 
-import net.bytebuddy.utility.RandomString
+import team.comit.simtong.domain.auth.model.value.Code
 import team.comit.simtong.global.DomainProperties.getProperty
 import team.comit.simtong.global.DomainPropertiesPrefix
 import team.comit.simtong.global.annotation.Aggregate
@@ -38,29 +38,5 @@ data class AuthCode(
             code = Code.defaultValue(),
             expirationTime = EXPIRED
         )
-    }
-}
-
-/**
- *
- * AuthCode Aggregate 중 인증코드를 담당하는 Code
- *
- * @author kimbeomjin
- * @date 2023/01/09
- * @version 1.2.5
- **/
-@JvmInline
-value class Code private constructor(
-    val value: String
-) {
-
-    fun match(code: String): Boolean {
-        return this.value == code
-    }
-
-    companion object {
-        fun of(value: String) = Code(value)
-
-        fun defaultValue() = Code(RandomString(6).nextString())
     }
 }
